+++ examples/semaphores/pr1/main.c
@@ -22,7 +22,7 @@
 uint8_t sid;
 
 int main() {
-  uint8_t tid;
+  uint32_t tid;
   pok_ret_t ret;
   pok_thread_attr_t tattr;
